Fixed a referencing error in SerializationUtilsTest.java
